fifth commit

+++ database/migrations/2024_11_11_222946_create_produits_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->timestamps();
             $table->string('nom');
             $table->float('qteProduit');
-            $table->float('prixProduit');
+            $table->decimal('prixProduit', 10, 2);
         });
     }
